feat : 방문기록에 따른 테이블 컴포넌트 렌더링
