+++ pages/patient/[id].js
@@ -1,34 +1,23 @@
-import React, { useEffect } from 'react'
+import React from 'react'
 import Link from 'next/link'
-
 import { useRouter } from 'next/router'
-import { useDispatch, useSelector } from 'react-redux'
+import { BiArrowBack } from 'react-icons/bi'
+import PatientDetailpageContainer from '@patient/containers/PatientDetailPageContainer'
 
 const Patient = () => {
   const router = useRouter()
   const { id } = router.query
 
-  const { patientCond, patientDrug, patientVisit } = useSelector(
-    (state) => state.api,
-  )
-  const dispatch = useDispatch()
-
-  useEffect(() => {
-    detailInfoFetch()
-  }, [])
-
-  const detailInfoFetch = () => {
-    dispatch({ type: 'FETCH_DATA', fetchType: 'patientCond', id })
-    dispatch({ type: 'FETCH_DATA', fetchType: 'patientDrug', id })
-    dispatch({ type: 'FETCH_DATA', fetchType: 'patientVisit', id })
-  }
-
   return (
-    <div>
+    <>
       <Link href="/">
-        <a>전체 환자 목록</a>
+        <div>
+          <BiArrowBack />
+          <a>전체 환자 목록</a>
+        </div>
       </Link>
-    </div>
+      {id && <PatientDetailpageContainer id={id} />}
+    </>
   )
 }
