+++ patient/consts/patientDetailConst.js
@@ -1,0 +1,6 @@
+export const pVisitCategory = [
+  { id: 'visitID', value: '방문 ID', tableCol: 'visitID' },
+  { id: 'visitConceptName', value: '방문 유형', tableCol: 'visitConceptName' },
+  { id: 'visitStartDate', value: '방문 시작일', tableCol: 'visitStartDate' },
+  { id: 'visitEndDate', value: '방문 종료일', tableCol: 'visitEndDate' },
+]
